Spect.NetworkNode (spect v0.1.0)

Join between Network and Node to allow for the same node database to be used for multiple networks. This table is populated with information about a node per this network

Types

node_name()

@type node_name() :: String.t() | nil

Name of the node on this network

t()

@type t() :: %Spect.NetworkNode{
  __meta__: term(),
  id: term(),
  name: node_name(),
  network: Spect.Network.t(),
  network_id: Spect.Network.network_id(),
  node: Spect.Network.Node.t(),
  node_id: Spect.Network.network_id()
}

join of network and node